# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GPL API",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/user/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "user login",
                "description": "user can login with credentials",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/user/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "user logout",
                "description": "user logout",
                "operationId": "Logout",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "user Profile",
                "description": "User Profile Details",
                "operationId": "User Profile",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LoginRequest": {
                "title": "Login request",
                "description": "Loginrequest body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email",
                        "type": "string",
                        "example": "myemail"
                    },
                    "password": {
                        "title": "password",
                        "description": "password",
                        "type": "string",
                        "example": "mypassword"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}